Feat  crud paralela

+++ src/Repositories/Scores/NotaRepository.php
@@ -52,9 +52,9 @@ class NotaRepository {
             $bindings[':student_class_id'] = $params['student_class_id'];
         }
 
-        if (isset($params['bimester_id'])) {
-            $conditions[] = 'n.bimestre_id = :bimester_id';
-            $bindings[':bimester_id'] = $params['bimester_id'];
+        if (isset($params['period_id'])) {
+            $conditions[] = 'n.periodo_id = :period_id';
+            $bindings[':period_id'] = $params['period_id'];
         }
         
         if (count($conditions) > 0) {
@@ -84,7 +84,7 @@ class NotaRepository {
                 SET 
                     uuid = :uuid,
                     atividade_id = :atividade_id,
-                    bimestre_id = :bimestre_id,
+                    periodo_id = :periodo_id,
                     estudante_turma_id = :estudante_turma_id,
                     nota = :nota"
             );
@@ -97,7 +97,7 @@ class NotaRepository {
             $create = $stmt->execute([
                 ':uuid' => $class->uuid,
                 ':atividade_id' => $class->atividade_id,
-                ':bimestre_id' => $class->bimestre_id,
+                ':periodo_id' => $class->periodo_id,
                 ':estudante_turma_id' => $class->estudante_turma_id,
                 ':nota' => $class->nota
             ]);
@@ -118,10 +118,10 @@ class NotaRepository {
 
     private function checkIfExistsScore($class) :?String {
         try {
-            $stmt = $this->conn->prepare("SELECT id FROM notas WHERE estudante_turma_id = :estudante_turma_id AND bimestre_id = :bimestre_id AND atividade_id = :atividade_id");
+            $stmt = $this->conn->prepare("SELECT id FROM notas WHERE estudante_turma_id = :estudante_turma_id AND periodo_id = :periodo_id AND atividade_id = :atividade_id");
             $stmt->execute([
                 ':estudante_turma_id' => $class->estudante_turma_id,
-                ':bimestre_id' => $class->bimestre_id,
+                ':periodo_id' => $class->periodo_id,
                 ':atividade_id' => $class->atividade_id
             ]);
 
@@ -137,7 +137,14 @@ class NotaRepository {
         }
     }
 
-    private function removeScore($id) :?bool {
+    private function removeScore($id) :?bool 
+    {
+        $scores = $this->findById((int)$id);
+
+        if (is_null($scores)) {
+            return null;
+        }
+        
         try {
             $stmt = $this->conn->prepare("DELETE FROM notas WHERE id = :id");
             $delete = $stmt->execute([
@@ -160,7 +167,7 @@ class NotaRepository {
     {
         $sql = "SELECT 
                     d.nome AS disciplina,
-                    b.bimestre AS bimestre,
+                    b.periodo AS periodo,
                     GROUP_CONCAT(CONCAT(a.tipo, ': ', n.nota) SEPARATOR '</br> ') AS atividades_notas,
                     pf.nome as professor
                 FROM 
@@ -180,7 +187,7 @@ class NotaRepository {
                 LEFT JOIN 
                     disciplinas d ON pd.disciplina_id = d.id
                 LEFT JOIN 
-                    bimestres b ON n.bimestre_id = b.id                
+                    periodo b ON n.periodo_id = b.id                
             ";
         
         
@@ -192,9 +199,9 @@ class NotaRepository {
             $bindings[':student_class_id'] = $params['student_class_id'];
         }
 
-        if (isset($params['bimester_id'])) {
-            $conditions[] = 'n.bimestre_id = :bimester_id';
-            $bindings[':bimester_id'] = $params['bimester_id'];
+        if (isset($params['period_id'])) {
+            $conditions[] = 'n.periodo_id = :period_id';
+            $bindings[':period_id'] = $params['period_id'];
         }
         
         if (count($conditions) > 0) {
@@ -204,7 +211,7 @@ class NotaRepository {
         $sql .= " GROUP BY 
                     pf.nome, et.id, d.id, b.id
                 ORDER BY 
-                 d.nome, b.bimestre;";
+                 d.nome, b.periodo;";
 
         try {
             $stmt = $this->conn->prepare($sql);
